Move achievement progress calcs to selector
