+++ app/src/redux/selectors/achievements.js
@@ -1,5 +1,7 @@
 import { createSelector } from 'reselect';
 import _ from 'lodash';
+import { getPlayer } from './players';
+import { getTotalLevel } from '../../utils';
 
 const achievementsSelector = state => state.achievements.achievements;
 
@@ -12,6 +14,7 @@ export const getAchievements = createSelector(achievementsSelector, map => Objec
 export const getPlayerAchievements = (state, playerId) => getAchievementsMap(state)[playerId];
 
 export const getPlayerAchievementsGrouped = (state, playerId) => {
+  const player = getPlayer(state, playerId);
   const list = getPlayerAchievements(state, playerId);
 
   if (!list) {
@@ -37,8 +40,49 @@ export const getPlayerAchievementsGrouped = (state, playerId) => {
     previousMeasure = s.measure;
   });
 
-  return groups;
+  const processed = groups
+    .map(group => processGroup(player, group))
+    .sort((a, b) => a.achievements.length - b.achievements.length);
+
+  return processed;
 };
+
+function processGroup(player, group) {
+  if (!player) {
+    return group;
+  }
+
+  const { latestSnapshot } = player;
+
+  if (group.metric === 'combat') {
+    const progress = player.combatLevel / 126;
+    return { ...group, achievements: [...group.achievements.map(a => ({ ...a, progress }))] };
+  }
+
+  if (group.metric === 'overall' && group.measure === 'levels') {
+    const progress = getTotalLevel(latestSnapshot) / 2277;
+    return { ...group, achievements: [...group.achievements.map(a => ({ ...a, progress }))] };
+  }
+
+  if (latestSnapshot[group.metric]) {
+    const currentValue = latestSnapshot[group.metric][group.measure];
+
+    const processedAchievements = group.achievements.map((achievement, i) => {
+      if (currentValue >= achievement.value) {
+        return { ...achievement, progress: 1 };
+      }
+
+      const prevStart = i === 0 ? 0 : group.achievements[i - 1].value;
+      const currentProgress = Math.max(0, (currentValue - prevStart) / (achievement.value - prevStart));
+
+      return { ...achievement, progress: currentProgress };
+    });
+
+    return { ...group, achievements: processedAchievements };
+  }
+
+  return group;
+}
 
 function formatAchievement(a) {
   const isDateUnknown = a.createdAt && a.createdAt.getFullYear() < 2000;
